refactor: rename `kcp scan connect clusters` to `kcp scan connect-topics`
Flatten the two-level `scan connect clusters` subcommand into a single
`scan connect-topics` command, since the only thing the `connect` group
contained was `clusters`. The command's purpose — parsing connect-status
topics to discover Connect workers — is more accurately reflected by the
`connect-topics` name.
