+++ cmd/scan/cmd_scan.go
@@ -3,7 +3,7 @@ package scan
 import (
 	"github.com/confluentinc/kcp/cmd/scan/client_inventory"
 	"github.com/confluentinc/kcp/cmd/scan/clusters"
-	"github.com/confluentinc/kcp/cmd/scan/connect"
+	connect_topics "github.com/confluentinc/kcp/cmd/scan/connect_topics"
 	"github.com/confluentinc/kcp/cmd/scan/schema_registry"
 	"github.com/confluentinc/kcp/cmd/scan/self_managed_connectors"
 	"github.com/spf13/cobra"
@@ -21,7 +21,7 @@ func NewScanCmd() *cobra.Command {
 	scanCmd.AddCommand(
 		client_inventory.NewScanClientInventoryCmd(),
 		clusters.NewScanClustersCmd(),
-		connect.NewScanConnectCmd(),
+		connect_topics.NewScanConnectTopicsCmd(),
 		schema_registry.NewScanSchemaRegistryCmd(),
 		self_managed_connectors.NewScanSelfManagedConnectorsCmd(),
 	)
